fix: マイグレーションファイルの型とべき等性の問題を修正
- task_dependencies の外部キーカラムを integer から bigint に修正
- 既存カラムチェックを追加してマイグレーションのべき等性を確保
- パフォーマンスインデックスをローカル環境に適用

+++ backend/db/migrate/20251217185831_create_task_dependencies.rb
@@ -1,8 +1,8 @@
 class CreateTaskDependencies < ActiveRecord::Migration[8.0]
   def change
     create_table :task_dependencies do |t|
-      t.integer :predecessor_id, null: false
-      t.integer :successor_id, null: false
+      t.bigint :predecessor_id, null: false
+      t.bigint :successor_id, null: false
 
       t.timestamps
     end
